fastapi_part/requirements.txt
numpy==1.24.3
# numpy==1.17.4
fastapi==0.111.0
# keras==3.3.3
matplotlib==3.7.2
# Pillow==10.1.0
# Pillow==7.0.0
Pillow==10.3.0
prometheus_client==0.17.1
prometheus_fastapi_instrumentator==7.0.0
pydantic==2.7.1
typing_extensions==4.11.0
uvicorn==0.29.0
efficientnet-pytorch
torch
albumentations
torchvision